Moved ParserSemanticPredicatesUtil, AcfKeywordHelper, and IgnoreCaseString to exported package com.avaloq.tools.ddk.xtext.generator.util.
